Writing client handshakes

+++ Makefile
@@ -1,13 +1,16 @@
+circd: $(wildcard src/*.[ch])
+	$(info :Compiling)
+	clang -o $@ \
+		-lpthread \
+		-Wall -Wextra -Wpedantic -O0 -g \
+		$(filter %.c, $^)
+
 .PHONY: run
 run: circd
 	$(info :Running)
 	./circd
 
-circd: $(wildcard src/*.[ch])
-	$(info :Compiling)
-	clang -o $@ -Wall -Wextra -Wpedantic -Werror -O0 -g $(filter %.c, $^)
-
 .PHONY: clean
 clean:
 	$(info :Cleaning)
-	rm circd
+	test -f circd && rm circd
